requirements_synthesis.txt
# Requirements for Data Synthesis Application

numpy>=1.20.0
scipy>=1.7.0
pillow>=9.0.0
opencv-python>=4.5.0
PyOpenGL>=3.1.5
glfw>=2.0.0
imgui[glfw]>=2.0.0
pyassimp>=4.1.4
plyfile>=0.7.4
